+++ lib/features/modeling/behavior/index.js
@@ -20,6 +20,7 @@ import ModelingFeedback from './ModelingFeedback';
 import ReplaceConnectionBehavior from './ReplaceConnectionBehavior';
 import RemoveParticipantBehavior from './RemoveParticipantBehavior';
 import ReplaceElementBehaviour from './ReplaceElementBehaviour';
+import ResizeBehavior from './ResizeBehavior';
 import ResizeLaneBehavior from './ResizeLaneBehavior';
 import RemoveElementBehavior from './RemoveElementBehavior';
 import SubProcessStartEventBehavior from './SubProcessStartEventBehavior';
@@ -53,6 +54,7 @@ export default {
     'removeParticipantBehavior',
     'replaceConnectionBehavior',
     'replaceElementBehaviour',
+    'resizeBehavior',
     'resizeLaneBehavior',
     'toggleElementCollapseBehaviour',
     'subProcessStartEventBehavior',
@@ -82,6 +84,7 @@ export default {
   replaceConnectionBehavior: [ 'type', ReplaceConnectionBehavior ],
   removeParticipantBehavior: [ 'type', RemoveParticipantBehavior ],
   replaceElementBehaviour: [ 'type', ReplaceElementBehaviour ],
+  resizeBehavior: [ 'type', ResizeBehavior ],
   resizeLaneBehavior: [ 'type', ResizeLaneBehavior ],
   removeElementBehavior: [ 'type', RemoveElementBehavior ],
   toggleElementCollapseBehaviour : [ 'type', ToggleElementCollapseBehaviour ],
