+++ lib/features/snapping/BpmnSnapping.js
@@ -11,8 +11,6 @@ import { is } from '../../util/ModelUtil';
 
 import { isAny } from '../modeling/util/ModelingUtil';
 
-import { isExpanded } from '../../util/DiUtil';
-
 import {
   bottomRight,
   isSnapped,
@@ -23,10 +21,7 @@ import {
 
 import { asTRBL } from 'diagram-js/lib/layout/LayoutUtil';
 
-import {
-  getBoundaryAttachment,
-  getParticipantSizeConstraints
-} from './BpmnSnappingUtil';
+import { getBoundaryAttachment } from './BpmnSnappingUtil';
 
 import { getLanesRoot } from '../modeling/util/LaneUtil';
 
@@ -150,33 +145,6 @@ export default function BpmnSnapping(bpmnRules, elementRegistry, eventBus, injec
     }
 
   });
-
-
-  eventBus.on('resize.start', HIGH_PRIORITY, function(event) {
-    var context = event.context,
-        shape = context.shape;
-
-    if (is(shape, 'bpmn:SubProcess') && isExpanded(shape)) {
-      context.minDimensions = { width: 140, height: 120 };
-    }
-
-    if (is(shape, 'bpmn:Participant')) {
-      context.minDimensions = { width: 300, height: 150 };
-    }
-
-    if (is(shape, 'bpmn:Lane') || is(shape, 'bpmn:Participant')) {
-      context.resizeConstraints = getParticipantSizeConstraints(
-        shape,
-        context.direction,
-        context.balanced
-      );
-    }
-
-    if (is(shape, 'bpmn:TextAnnotation')) {
-      context.minDimensions = { width: 50, height: 30 };
-    }
-  });
-
 }
 
 inherits(BpmnSnapping, Snapping);
@@ -260,7 +228,6 @@ BpmnSnapping.prototype.initSnap = function(event) {
     snapContext.addDefaultSnap('mid', mid(source));
   }
 };
-
 
 BpmnSnapping.prototype.addTargetSnaps = function(snapPoints, shape, target) {
