+++ test/spec/features/modeling/behavior/ResizeBehavior.lanes.bpmn
