+++ test/spec/features/modeling/behavior/ResizeBehavior.participant.bpmn
