+++ test/spec/features/modeling/behavior/ResizeBehavior.subProcess.bpmn
