+++ test/spec/features/modeling/behavior/ResizeBehavior.textAnnotation.bpmn
