+++ test/spec/features/modeling/behavior/util/ResizeUtil.lanes.bpmn
