+++ test/spec/features/snapping/BpmnSnappingSpec.js
@@ -13,7 +13,6 @@ import coreModule from 'lib/core';
 import snappingModule from 'lib/features/snapping';
 import modelingModule from 'lib/features/modeling';
 import createModule from 'diagram-js/lib/features/create';
-import resizeModule from 'diagram-js/lib/features/resize';
 import moveModule from 'diagram-js/lib/features/move';
 import rulesModule from 'lib/features/rules';
 import connectModule from 'diagram-js/lib/features/connect';
@@ -449,215 +448,12 @@ describe('features/snapping - BpmnSnapping', function() {
   });
 
 
-  describe('on Participant resize', function() {
-
-    describe('snap min bounds', function() {
-
-      var diagramXML = require('./BpmnSnapping.participant-resize.bpmn');
-
-      var testResizeModules = [
-        coreModule,
-        resizeModule,
-        modelingModule,
-        rulesModule,
-        snappingModule
-      ];
-
-      beforeEach(bootstrapModeler(diagramXML, { modules: testResizeModules }));
-
-
-      it('should snap to children from <se>', inject(function(elementRegistry, resize, dragging) {
-
-        var participant = elementRegistry.get('Participant_2');
-
-        resize.activate(canvasEvent({ x: 500, y: 500 }), participant, 'se');
-        dragging.move(canvasEvent({ x: 0, y: 0 }));
-        dragging.end();
-
-        expect(participant.width).to.equal(482);
-        expect(participant.height).to.equal(252);
-      }));
-
-
-      it('should snap to children from <nw>', inject(function(elementRegistry, resize, dragging) {
-
-        var participant = elementRegistry.get('Participant_2');
-
-        resize.activate(canvasEvent({ x: 0, y: 0 }), participant, 'nw');
-        dragging.move(canvasEvent({ x: 500, y: 500 }));
-        dragging.end();
-
-        expect(participant.width).to.equal(467);
-        expect(participant.height).to.equal(287);
-      }));
-
-
-      it('should snap to min dimensions from <se>', inject(function(elementRegistry, resize, dragging) {
-
-        var participant = elementRegistry.get('Participant_1');
-
-        resize.activate(canvasEvent({ x: 500, y: 500 }), participant, 'se');
-        dragging.move(canvasEvent({ x: 0, y: 0 }));
-        dragging.end();
-
-        expect(participant.width).to.equal(300);
-        expect(participant.height).to.equal(60);
-      }));
-
-
-      it('should snap to min dimensions from <nw>', inject(function(elementRegistry, resize, dragging) {
-
-        var participant = elementRegistry.get('Participant_1');
-
-        resize.activate(canvasEvent({ x: 0, y: 0 }), participant, 'nw');
-        dragging.move(canvasEvent({ x: 500, y: 500 }));
-        dragging.end();
-
-        expect(participant.width).to.equal(300);
-        expect(participant.height).to.equal(60);
-      }));
-
-
-      it('should snap to min dimensions + children from <se>', inject(function(elementRegistry, resize, dragging) {
-
-        var participant = elementRegistry.get('Participant_3');
-
-        resize.activate(canvasEvent({ x: 500, y: 500 }), participant, 'se');
-        dragging.move(canvasEvent({ x: 0, y: 0 }));
-        dragging.end();
-
-        expect(participant.width).to.equal(305);
-
-        // snap to children rather than min dimensions
-        expect(participant.height).to.equal(143);
-      }));
-
-
-      it('should snap to min dimensions + children from <nw>', inject(function(elementRegistry, resize, dragging) {
-
-        var participant = elementRegistry.get('Participant_3');
-
-        resize.activate(canvasEvent({ x: 0, y: 0 }), participant, 'nw');
-        dragging.move(canvasEvent({ x: 500, y: 500 }));
-        dragging.end();
-
-        expect(participant.width).to.equal(353);
-        expect(participant.height).to.equal(177);
-      }));
-
-    });
-
-
-    describe('snap child lanes', function() {
-
-      var diagramXML = require('./BpmnSnapping.lanes-resize.bpmn');
-
-      var testResizeModules = [
-        coreModule,
-        resizeModule,
-        modelingModule,
-        rulesModule,
-        snappingModule
-      ];
-
-      beforeEach(bootstrapModeler(diagramXML, { modules: testResizeModules }));
-
-
-      it('should snap to child lanes from <nw>', inject(function(elementRegistry, resize, dragging) {
-
-        var participant = elementRegistry.get('Participant');
-
-        resize.activate(canvasEvent({ x: 0, y: 0 }), participant, 'nw');
-        dragging.move(canvasEvent({ x: 500, y: 500 }));
-        dragging.end();
-
-        expect(participant.width).to.equal(563);
-        expect(participant.height).to.equal(223);
-      }));
-
-
-      it('should snap to nested child lanes from <se>', inject(function(elementRegistry, resize, dragging) {
-
-        var lane = elementRegistry.get('Lane_B_0');
-
-        resize.activate(canvasEvent({ x: 0, y: 0 }), lane, 'se');
-        dragging.move(canvasEvent({ x: -500, y: -500 }));
-        dragging.end();
-
-        expect(lane.width).to.equal(313);
-        expect(lane.height).to.equal(122);
-      }));
-
-    });
-
-  });
-
-
-  describe('on SubProcess resize', function() {
-
-    var diagramXML = require('./BpmnSnapping.subProcess-resize.bpmn');
-
-    var testResizeModules = [
-      coreModule,
-      modelingModule,
-      resizeModule,
-      rulesModule,
-      snappingModule
-    ];
-
-    beforeEach(bootstrapModeler(diagramXML, { modules: testResizeModules }));
-
-
-    it('should snap to minimum bounds', inject(function(elementRegistry, resize, dragging) {
-
-      var subProcess = elementRegistry.get('SubProcess');
-
-      resize.activate(canvasEvent({ x: 0, y: 0 }), subProcess, 'se');
-      dragging.move(canvasEvent({ x: -400, y: -400 }));
-      dragging.end();
-
-      expect(subProcess.width).to.equal(140);
-      expect(subProcess.height).to.equal(120);
-    }));
-
-  });
-
-
-  describe('on TextAnnotation resize', function() {
-
-    var diagramXML = require('./BpmnSnapping.textAnnotation-resize.bpmn');
-
-    var testResizeModules = [
-      coreModule,
-      modelingModule,
-      resizeModule,
-      rulesModule,
-      snappingModule
-    ];
-
-    beforeEach(bootstrapModeler(diagramXML, { modules: testResizeModules }));
-
-
-    it('should snap to minimum bounds', inject(function(elementRegistry, resize, dragging) {
-
-      var textAnnotation = elementRegistry.get('TextAnnotation');
-
-      resize.activate(canvasEvent({ x: 0, y: 0 }), textAnnotation, 'se');
-      dragging.move(canvasEvent({ x: -400, y: -400 }));
-      dragging.end();
-
-      expect(textAnnotation.width).to.equal(50);
-      expect(textAnnotation.height).to.equal(30);
-    }));
-
-  });
-
-
   describe('labels', function() {
 
     var diagramXML = require('./BpmnSnapping.labels.bpmn');
 
     beforeEach(bootstrapModeler(diagramXML, { modules: testModules }));
+
 
     it('should snap to start events', inject(function(canvas, elementRegistry, move, dragging) {
